Reverting minmax_element test according to #1224 (review)
