--- test/parallel_api/algorithm/alg.sorting/alg.min.max/minmax_element_coal.pass.cpp
@@ -1,23 +1,0 @@
-// -*- C++ -*-
-//===-- minmax_element_coal.pass.cpp --------------------------------------===//
-//
-// Copyright (C) Intel Corporation
-//
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-// This file incorporates work covered by the following copyright and permission
-// notice:
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-//
-//===----------------------------------------------------------------------===//
-
-#define _ONEDPL_DETECT_SPIRV_COMPILATION 0
-#include "minmax_element.pass.h"
-
-int
-main()
-{
-    return run_tests();
-}
